+++ Sample/Sample.Web/ApiControllers/RolesController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Web.Http;
+using System.Web.Http.Description;
 using FluentValidation;
 using FluentValidation.Results;
 using Sample.Core.Models;
@@ -45,22 +46,22 @@ namespace Sample.Web.ApiControllers
 		}
 
 		// GET: api/role/5
-		[HttpGet, Route("{id}")]
-		public Role Get(int id)
+		[HttpGet, Route("{id}"), ResponseType(typeof(Role))]
+		public IHttpActionResult Get(int id)
 		{
 			Role role = _service.Get(id);
 			
 			if (role == null)
 			{
-			
+				return NotFound();
 			}
 
-			return role;
+			return Ok(role);
 		}
 
 		// POST: api/role
 		[HttpPost, Route("")]
-		public void Post([FromBody]Role role)
+		public IHttpActionResult Post([FromBody]Role role)
 		{
 			ValidationResult vr = _validator.Validate(role);
 
@@ -68,17 +69,22 @@ namespace Sample.Web.ApiControllers
 			{
 				_service.Insert(role);
 
-				return;
+				return Ok();
 			}
 			
-			throw new NotImplementedException();
+			return BadRequest(vr);
 		}
 
 		// PUT: api/role/5
 		[HttpPut, Route("{id}")]
-		public void Put(int id, [FromBody]Role role)
+		public IHttpActionResult Put(int id, [FromBody]Role role)
 		{
 			Role model = _service.Get(id);
+
+			if (model == null)
+			{
+				return NotFound();
+			}
 			
 			model.Name = role.Name;
 			model.CreatedAt = role.CreatedAt;
@@ -90,19 +96,40 @@ namespace Sample.Web.ApiControllers
 			{
 				_service.Update(model);
 
-				return;
+				return Ok();
 			}
 			
-			throw new NotImplementedException();
+			return BadRequest(vr);
 		}
 
 		// DELETE: api/role/5
 		[HttpDelete, Route("{id}")]
-		public void Delete(int id)
+		public IHttpActionResult Delete(int id)
 		{
 			Role model = _service.Get(id);
 
+			if (model == null)
+			{
+				return NotFound();
+			}
+
 			_service.Delete(model);
+			
+			return Ok();
+		}
+		
+		#endregion
+
+		#region Helpers
+
+		private IHttpActionResult BadRequest(ValidationResult vr)
+		{
+			foreach (ValidationFailure vf in vr.Errors)
+			{
+				ModelState.AddModelError(vf.PropertyName, vf.ErrorMessage);
+			}
+		
+			return BadRequest(ModelState);
 		}
 		
 		#endregion
